+++ lib/chewie.dart
@@ -1,3 +1,4 @@
 library chewie;
 
 export 'src/chewie_player.dart';
+export 'src/chewie_progress_colors.dart';