+++ lib/src/player_with_controls.dart
@@ -1,3 +1,4 @@
+import 'package:chewie/src/chewie_progress_colors.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:video_player/video_player.dart';
@@ -12,20 +13,26 @@ class PlayerWithControls extends StatefulWidget {
   final VideoPlayerController controller;
   final Future<dynamic> Function() onExpandCollapse;
   final bool fullScreen;
-  final VideoProgressColors progressColors;
+  final ChewieProgressColors cupertinoProgressColors;
+  final ChewieProgressColors materialProgressColors;
   final Widget placeholder;
   final double aspectRatio;
   final bool autoPlay;
+  final bool showControls;
 
   PlayerWithControls({
+    Key key,
     @required this.controller,
     @required this.onExpandCollapse,
     @required this.aspectRatio,
     this.fullScreen = false,
-    this.progressColors,
+    this.showControls = true,
+    this.cupertinoProgressColors,
+    this.materialProgressColors,
     this.placeholder,
     this.autoPlay,
-  });
+  })
+      : super(key: key);
 
   @override
   State createState() {
@@ -49,35 +56,43 @@ class _VideoPlayerWithControlsState extends State<PlayerWithControls> {
                 widget.placeholder ?? new Container(),
                 new Hero(
                   tag: controller,
-                  child: new Center(
-                    child: new AspectRatio(
-                        aspectRatio: widget.aspectRatio,
-                        child: new VideoPlayer(controller)),
+                  child: new AspectRatio(
+                    aspectRatio: widget.aspectRatio,
+                    child: new VideoPlayer(controller),
                   ),
                 ),
-                Theme.of(context).platform == TargetPlatform.android
-                    ? new MaterialControls(
-                        controller: controller,
-                        onExpandCollapse: widget.onExpandCollapse,
-                        fullScreen: widget.fullScreen,
-                        progressColors: widget.progressColors,
-                        autoPlay: widget.autoPlay,
-                      )
-                    : new CupertinoControls(
-                        backgroundColor: new Color.fromRGBO(41, 41, 41, 0.7),
-                        iconColor: new Color.fromARGB(255, 200, 200, 200),
-                        controller: controller,
-                        onExpandCollapse: widget.onExpandCollapse,
-                        fullScreen: widget.fullScreen,
-                        progressColors: widget.progressColors,
-                        autoPlay: widget.autoPlay,
-                      ),
+                _buildControls(context, controller),
               ],
             ),
           ),
         ),
       ),
     );
+  }
+
+  Widget _buildControls(
+    BuildContext context,
+    VideoPlayerController controller,
+  ) {
+    return widget.showControls
+        ? Theme.of(context).platform == TargetPlatform.android
+            ? new MaterialControls(
+                controller: controller,
+                onExpandCollapse: widget.onExpandCollapse,
+                fullScreen: widget.fullScreen,
+                progressColors: widget.materialProgressColors,
+                autoPlay: widget.autoPlay,
+              )
+            : new CupertinoControls(
+                backgroundColor: new Color.fromRGBO(41, 41, 41, 0.7),
+                iconColor: new Color.fromARGB(255, 200, 200, 200),
+                controller: controller,
+                onExpandCollapse: widget.onExpandCollapse,
+                fullScreen: widget.fullScreen,
+                progressColors: widget.cupertinoProgressColors,
+                autoPlay: widget.autoPlay,
+              )
+        : new Container();
   }
 
   @override